feat: reuseable stat badges for pokemon
