+++ src/components/PokemonTable.tsx
@@ -1,20 +1,8 @@
 "use client";
-import {
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  HeartIcon,
-  LoaderIcon,
-  PieChartIcon,
-  ShieldIcon,
-  ShieldPlusIcon,
-  SwordIcon,
-  SwordsIcon,
-  ZapIcon,
-} from "lucide-react";
+import { ChevronLeftIcon, ChevronRightIcon, LoaderIcon } from "lucide-react";
 import Image from "next/image";
 
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import {
   Pagination,
@@ -32,42 +20,12 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
 
 import { usePaginatedPokemonList } from "@/hooks/usePaginatedPokemonList";
 import { usePokemon } from "@/hooks/usePokemon";
 import { extractOffSetAndLimit } from "@/lib/utils";
 import { usePaginationContext } from "./PaginationContext";
-
-function StatIcon(statName: string) {
-  switch (statName) {
-    case "hp":
-      return <HeartIcon className="size-4" />;
-
-    case "attack":
-      return <SwordIcon className="size-4" />;
-
-    case "defense":
-      return <ShieldIcon className="size-4" />;
-
-    case "special-attack":
-      return <SwordsIcon className="size-4" />;
-
-    case "special-defense":
-      return <ShieldPlusIcon className="size-4" />;
-
-    case "speed":
-      return <ZapIcon className="size-4" />;
-
-    default:
-      return <PieChartIcon className="size-4" />;
-  }
-}
+import PokemonStatBadges from "./PokemonStatBadges";
 
 function LoadingRow() {
   return (
@@ -150,21 +108,7 @@ function PokemonRow({ name }: { name: string }) {
         </p>
       </TableCell>
       <TableCell className="grid grid-cols-2 md:grid-cols-3 place-items-center w-auto gap-[1px]">
-        {pokemon.stats.map((item) => (
-          <TooltipProvider key={item.stat.name}>
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <Badge className="cursor-default">
-                  {StatIcon(item.stat.name)}
-                  <p>={item.base_stat}</p>
-                </Badge>
-              </TooltipTrigger>
-              <TooltipContent>
-                <p className="capitalize">{item.stat.name.replace("-", " ")}</p>
-              </TooltipContent>
-            </Tooltip>
-          </TooltipProvider>
-        ))}
+        <PokemonStatBadges stats={pokemon.stats} />
       </TableCell>
       <TableCell className="">
         <p className="capitalize">
